+++ test-framework/src/test/java/uk/co/real_logic/artio/acceptance_tests/FakeAcceptanceTestHandler.java
@@ -37,7 +37,7 @@ public class FakeAcceptanceTestHandler extends FakeHandler
         final int libraryId,
         final Session session,
         final int sequenceIndex,
-        final int messageType,
+        final long messageType,
         final long timestampInNs,
         final long position)
     {
